financeiro_app/requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
